+++ argos/inspector/pgplugins/old_imageplot2d.py
@@ -177,6 +177,11 @@ class PgImagePlot2dCti(MainGroupCti):
 
         self.insertChild(PgGradientEditorItemCti(self.pgImagePlot2d.histLutItem.gradient))
 
+        # If True, the image is automatically downsampled to match the screen resolution. This
+        # improves performance for large images and reduces aliasing. If autoDownsample is not
+        # specified, then ImageItem will choose whether to downsample the image based on its size.
+        self.autoDownSampleCti = self.insertChild(BoolCti('auto down sample', True))
+
         self.zoomModeCti = self.insertChild(BoolCti('rectangle zoom mode', False))
 
         # Probe and cross-hair plots
@@ -502,6 +507,8 @@ class PgImagePlot2d(AbstractInspector):
         # We need to transpose the slicedArray ourselves because axes = {'x':1, 'y':0}
         # doesn't seem to do anything.
         imageArray = imageArray.transpose()
+
+        self.imageItem.setAutoDownsample(self.config.autoDownSampleCti.configValue)
         self.imageItem.setImage(imageArray, autoLevels=False)
 
         self.imagePlotItem.setRectangleZoomOn(self.config.zoomModeCti.configValue)